apps/datafeed/requirements.txt
# Core framework
Django>=5.0,<6.0
djangorestframework>=3.15.0
yfinance==0.2.66
celery==5.5.3
python-decouple==3.8
psycopg2-binary==2.9.10